Updated README.
Changed low power mode in demo access point from LPM3 to LPM0. The reason for this change deals (possibly) with the DCO used for the baud rate for the uart. Since the DCO is turned off during LPM3, the uart timings get screwed up after several transitions to-from LPM3. LPM0 does not turn off the DCO and the uart problems went away.

Reduced the packet size limit in RfRegSettings to 63. Currently the radio is operating with packets smaller than the 64-byte FIFO buffer. (See SLAA465 document for more information.) If for some reason, a packet larger than that is received, the buffer overflows and the radio stops working. The smaller packet limit discards any received packets larger than the limit.

In the future, the limit can be set to the maximum of 256, but the radio interface code needs to be modified.